Work trace

+++ core/vm/interface.go
@@ -41,7 +41,7 @@ type StateDB interface {
 	GetCodeHash(common.Address) common.Hash
 	GetCode(common.Address) []byte
 	SetCode(common.Address, []byte)
-	GetCodeSize(common.Address) int
+	GetCodeSize(common.Address) uint64
 
 	AddRefund(uint64)
 	GetRefund() uint64
